santiye-ai/backend/requirements.txt
fastapi
uvicorn
pydantic
openai
supabase
python-dotenv
requests
